Interfaces, static block,method and variable, final class,method and variable.
