+++ src/com/airport/RunwayStatus.java
@@ -1,0 +1,8 @@
+package com.airport;
+
+public enum RunwayStatus {
+    IDLE,
+    TAKEOFF,
+    LANDING
+}
+
