+++ src/com/person/Address.java
@@ -1,0 +1,62 @@
+package com.person;
+
+public class Address {
+    private String street;
+    private String city;
+    private String state;
+    private String zipcode;
+    private String country;
+
+    /*constructor*/
+
+    public Address(String street, String city, String state, String zipcode, String country) {
+        this.street = street;
+        this.city = city;
+        this.state = state;
+        this.zipcode = zipcode;
+        this.country = country;
+    }
+
+    public String getStreet() {
+        return street;
+    }
+
+    public void setStreet(String street) {
+        this.street = street;
+    }
+
+    public String getCity() {
+        return city;
+    }
+
+    public void setCity(String city) {
+        this.city = city;
+    }
+    public String getState() {
+        return state;
+    }
+    public void setState(String state) {
+        this.state = state;
+    }
+    public String getCountry() {
+        return country;
+    }
+    public void setCountry(String country) {
+        this.country = country;
+    }
+    public String getZipcode() {
+        return zipcode;
+    }
+    public void setZipcode(String zipcode) {
+        this.zipcode = zipcode;
+    }
+    public String getaddress() {
+        return "Address{" +
+                "street:" + street +
+                "city :" + city +
+                "state :" + state +
+                "country :" + country +
+                "zipcode :" + zipcode +
+                "}";
+    }
+}
