+++ src/com/person/Designation.java
@@ -1,0 +1,9 @@
+package com.person;
+
+public enum Designation {
+    PILOT,
+    AIRHOSTESS,
+    CARGO_MANAGER,
+    CUSTOMERCARE,
+    MANAGER;
+}
