+++ src/com/person/Gender.java
@@ -1,0 +1,7 @@
+package com.person;
+
+public enum Gender {
+    MALE,
+    FEMALE,
+    NOT_DISCLOSED;
+}
