+++ src/com/person/Item.java
@@ -1,0 +1,10 @@
+package com.person;
+
+public interface Item {
+    void addItem(String item);
+    void removeItem(String item);
+    Boolean hasMetalObject();
+    Boolean hasExplosiveObject();
+
+
+}
