+++ src/com/utils/Printer.java
@@ -1,0 +1,11 @@
+package com.utils;
+
+public class Printer {
+    public static void print(String out) {
+        System.out.println("AIRPORT: " + out);
+    }
+
+    public static void print(int out) {
+        System.out.println("AIRPORT: " + out);
+    }
+}
