--- src/task2/airport/RunwayStatus.java
@@ -1,6 +1,0 @@
-package task2.airport;
-
-public enum RunwayStatus {
-    // Runway status enum
-    IDLE, LANDING, TAKEOFF;
-}
