--- src/task2/person/Designation.java
@@ -1,5 +1,0 @@
-package task2.person;
-
-public enum Designation {
-    PILOT, AIRHOSTESS, CARGO_MANAGER, CUSTOMERCARE, MANAGER;
-}
