--- src/task2/person/Gender.java
@@ -1,5 +1,0 @@
-package task2.person;
-
-public enum Gender {
-    MALE, FEMALE, NOT_DISCLOSED;
-}
