--- src/task2/utils/Printer.java
@@ -1,7 +1,0 @@
-package task2.utils;
-
-public class Printer {
-    public static void print(String out) {
-        System.out.println("AIRPORT: " + out);
-    }
-}
